requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
autopep8==2.1.0
billiard==4.2.0
cachetools==5.3.3
celery==5.4.0
celery[redis]==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.7.1
django-cors-headers==4.3.1
django-filter==24.2
django-silk==0.5.2
django-templated-mail==1.1.1
django==5.0.4
djangorestframework-simplejwt==5.3.1
djangorestframework==3.15.1
djoser==2.2.2
drf-yasg[validation]==1.21.7
faker==24.11.0
flower==2.0.1
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth-httplib2==0.2.0
google-auth==2.29.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
humanize==4.9.0
idna==3.7
inflection==0.5.1
jinja2==3.1.3
jsonschema-specifications==2023.12.1
jsonschema==4.21.1
kombu==5.3.7
markupsafe==2.1.5
oauthlib==3.2.2
packaging==24.0
prometheus-client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
pyasn1-modules==0.4.0
pyasn1==0.6.0
pycodestyle==2.11.1
pycparser==2.22
pygments==2.17.2
pyjwt==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2024.1
pyyaml==6.0.1
redis==5.0.3
referencing==0.34.0
requests-oauthlib==2.0.0
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
setuptools==69.5.1
simplejson==3.19.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.5.0
swagger-spec-validator==3.0.3
tornado==6.4
typing-extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13